first attempt to extract common parts of two main files
