+++ official/resnet/keras/keras_cifar_main.py
@@ -27,56 +27,12 @@ import tensorflow as tf  # pylint: disable=g-bad-import-order
 
 from official.resnet import cifar10_main as cifar_main
 from official.resnet import resnet_run_loop
+from official.resnet.keras import keras_common
 from official.resnet.keras import keras_resnet_model
 from official.utils.flags import core as flags_core
 from official.utils.logs import logger
 from official.utils.misc import distribution_utils
 from tensorflow.python.keras.optimizer_v2 import gradient_descent as gradient_descent_v2
-
-
-class TimeHistory(tf.keras.callbacks.Callback):
-  """Callback for Keras models."""
-
-  def __init__(self, batch_size):
-    """Callback for Keras models.
-
-    Args:
-      batch_size: Total batch size.
-
-    """
-    self._batch_size = batch_size
-    self.last_exp_per_sec = 0
-    super(TimeHistory, self).__init__()
-
-  def on_train_begin(self, logs=None):
-    self.epoch_times_secs = []
-    self.batch_times_secs = []
-    self.record_batch = True
-
-  def on_epoch_begin(self, epoch, logs=None):
-    self.epoch_time_start = time.time()
-
-  def on_epoch_end(self, epoch, logs=None):
-    self.epoch_times_secs.append(time.time() - self.epoch_time_start)
-
-  def on_batch_begin(self, batch, logs=None):
-    if self.record_batch:
-      self.batch_time_start = time.time()
-      self.record_batch = False
-
-  def on_batch_end(self, batch, logs=None):
-    n = 100
-    if batch % n == 0:
-      last_n_batches = time.time() - self.batch_time_start
-      examples_per_second = (self._batch_size * n) / last_n_batches
-      self.batch_times_secs.append(last_n_batches)
-      self.last_exp_per_sec = examples_per_second
-      self.record_batch = True
-      # TODO(anjalisridhar): add timestamp as well.
-      if batch != 0:
-        tf.logging.info("BenchmarkMetric: {'num_batches':%d, 'time_taken': %f,"
-                        "'images_per_second': %f}" %
-                        (batch, last_n_batches, examples_per_second))
 
 
 # LR_SCHEDULE = [    # (multiplier, epoch to start) tuples
@@ -125,42 +81,6 @@ def learning_rate_schedule(current_epoch, current_batch, batches_per_epoch, batc
     else:
       break
   return learning_rate
-
-
-class LearningRateBatchScheduler(tf.keras.callbacks.Callback):
-  """Callback to update learning rate on every batch (not epoch boundaries).
-
-  N.B. Only support Keras optimizers, not TF optimizers.
-
-  Args:
-      schedule: a function that takes an epoch index and a batch index as input
-          (both integer, indexed from 0) and returns a new learning rate as
-          output (float).
-  """
-
-  def __init__(self, schedule, batch_size, num_images):
-    super(LearningRateBatchScheduler, self).__init__()
-    self.schedule = schedule
-    self.batches_per_epoch = num_images / batch_size
-    self.batch_size = batch_size
-    self.epochs = -1
-    self.prev_lr = -1
-
-  def on_epoch_begin(self, epoch, logs=None):
-    #if not hasattr(self.model.optimizer, 'learning_rate'):
-    #  raise ValueError('Optimizer must have a "learning_rate" attribute.')
-    self.epochs += 1
-
-  def on_batch_begin(self, batch, logs=None):
-    lr = self.schedule(self.epochs, batch, self.batches_per_epoch, self.batch_size)
-    if not isinstance(lr, (float, np.float32, np.float64)):
-      raise ValueError('The output of the "schedule" function should be float.')
-    if lr != self.prev_lr:
-      tf.keras.backend.set_value(self.model.optimizer.learning_rate, lr)
-      self.prev_lr = lr
-      tf.logging.debug('Epoch %05d Batch %05d: LearningRateBatchScheduler change '
-                   'learning rate to %s.', self.epochs, batch, lr)
-
 
 
 def parse_record_keras(raw_record, is_training, dtype):
@@ -244,56 +164,23 @@ def run_cifar_with_keras(flags_obj):
         num_epochs=flags_obj.train_epochs,
         parse_record_fn=parse_record_keras)
 
-  # Use Keras ResNet50 applications model and native keras APIs
-  # initialize RMSprop optimizer
-  # TODO(anjalisridhar): Move to using MomentumOptimizer.
-  # opt = tf.train.GradientDescentOptimizer(learning_rate=0.0001)
-  # I am setting an initial LR of 0.001 since this will be reset
-  # at the beginning of the training loop.
-  opt = gradient_descent_v2.SGD(learning_rate=0.1, momentum=0.9)
-
-  # TF Optimizer:
-  # opt = tf.train.MomentumOptimizer(learning_rate=0.1, momentum=0.9)
-
-  strategy = distribution_utils.get_distribution_strategy(
-      num_gpus=flags_obj.num_gpus)
+  opt, loss, accuracy = keras_common.get_optimizer_loss_and_metrics()
+  strategy = keras_common.get_dist_strategy()
 
   model = keras_resnet_model.ResNet56(input_shape=(32, 32, 3),
                                       include_top=True,
                                       classes=cifar_main._NUM_CLASSES,
                                       weights=None)
-
-  loss = 'categorical_crossentropy'
-  accuracy = 'categorical_accuracy'
-
-  if flags_obj.num_gpus == 1 and flags_obj.dist_strat_off:
-    print('Not using distribution strategies.')
-    model.compile(loss=loss,
-                  optimizer=opt,
-                  metrics=[accuracy])
-  else:
-    model.compile(loss=loss,
-                  optimizer=opt,
-                  metrics=[accuracy],
-                  distribute=strategy)
+  model.compile(loss=loss,
+                optimizer=opt,
+                metrics=[accuracy],
+                distribute=strategy)
+  time_callback, tensorboard_callback, lr_callback = keras_common.get_fit_callbacks()
 
   steps_per_epoch = cifar_main._NUM_IMAGES['train'] // flags_obj.batch_size
-
-  time_callback = TimeHistory(flags_obj.batch_size)
-
-  tesorboard_callback = tf.keras.callbacks.TensorBoard(
-      log_dir=flags_obj.model_dir)
-    #  update_freq="batch")  # Add this if want per batch logging.
-
-  lr_callback = LearningRateBatchScheduler(
-      learning_rate_schedule,
-      batch_size=flags_obj.batch_size,
-      num_images=cifar_main._NUM_IMAGES['train'])
-
   num_eval_steps = (cifar_main._NUM_IMAGES['validation'] //
                     flags_obj.batch_size)
 
-  print('Executing eagerly?:', tf.executing_eagerly())
   history = model.fit(train_input_dataset,
                       epochs=flags_obj.train_epochs,
                       steps_per_epoch=steps_per_epoch,
@@ -310,14 +197,9 @@ def run_cifar_with_keras(flags_obj):
                                steps=num_eval_steps,
                                verbose=1)
 
-  stats = {}
-  stats['accuracy_top_1'] = eval_output[1]
-  stats['eval_loss'] = eval_output[0]
-  stats['training_loss'] = history.history['loss'][-1]
-  stats['training_accuracy_top_1'] = history.history['categorical_accuracy'][-1]
+  print('Test loss:', eval_output[0])
+  stats = keras_common.analyze_eval_result(eval_output)
 
-  print('top_1 accuracy:{}'.format(stats['accuracy_top_1']))
-  print('top_1_training_accuracy:{}'.format(stats['training_accuracy_top_1']))
   return stats
